worked on issue #4, border line mode, with partially now working, but not constructing longer lines than two pixels

+++ src/outlinersvg.hh
@@ -24,7 +24,7 @@ class SvgCreator {
              float yStart,
              float xFactor,
              float yFactor,
-             unsigned int linewidth);
+             float linewidth);
   ~SvgCreator();
   void line(float fromX,
             float fromY,
@@ -43,7 +43,7 @@ class SvgCreator {
   float yStart;
   float xFactor;
   float yFactor;
-  unsigned int linewidth;
+  float linewidth;
 
   void preamble();
   void postamble();
